# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ShopEasy e-commerce API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LoginInput {
  password: string;
}

export interface AuthResponse {
  authenticated: boolean;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  /** @nullable */
  icon?: string | null;
  /** @nullable */
  productCount?: number | null;
}

export interface CategoryInput {
  name: string;
  slug: string;
  /** @nullable */
  icon?: string | null;
}

export interface Product {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  price: number;
  offerPrice: number;
  /** @nullable */
  discountPercent?: number | null;
  images: string[];
  /** @nullable */
  categoryId?: number | null;
  /** @nullable */
  categoryName?: string | null;
  inStock: boolean;
  featured?: boolean;
  /** @nullable */
  stockCount?: number | null;
  createdAt: string;
}

export interface ProductInput {
  name: string;
  /** @nullable */
  description?: string | null;
  price: number;
  offerPrice: number;
  images?: string[];
  /** @nullable */
  categoryId?: number | null;
  inStock?: boolean;
  featured?: boolean;
  /** @nullable */
  stockCount?: number | null;
}

export interface StockUpdate {
  inStock: boolean;
  /** @nullable */
  stockCount?: number | null;
}

export interface OrderItem {
  productId: number;
  productName: string;
  /** @nullable */
  productImage?: string | null;
  quantity: number;
  price: number;
  offerPrice: number;
}

export interface Order {
  id: number;
  orderId: string;
  customerName: string;
  phone: string;
  address: string;
  items: OrderItem[];
  total: number;
  paymentMethod: string;
  status: string;
  createdAt: string;
}

export interface OrderInput {
  customerName: string;
  phone: string;
  address: string;
  items: OrderItem[];
  total: number;
  paymentMethod: string;
}

export interface OrderStatusUpdate {
  status: string;
}

export interface OrderStats {
  total: number;
  pending: number;
  processing: number;
  shipped: number;
  delivered: number;
  cancelled: number;
  totalRevenue: number;
  todayOrders: number;
  codOrders: number;
  onlineOrders: number;
}

export interface Banner {
  id: number;
  title: string;
  /** @nullable */
  subtitle?: string | null;
  imageUrl: string;
  /** @nullable */
  linkUrl?: string | null;
  active: boolean;
  sortOrder: number;
}

export interface BannerInput {
  title: string;
  /** @nullable */
  subtitle?: string | null;
  imageUrl: string;
  /** @nullable */
  linkUrl?: string | null;
  active?: boolean;
  sortOrder?: number;
}

export interface Settings {
  storeName: string;
  /** @nullable */
  storeTagline?: string | null;
  whatsappNumber: string;
  /** @nullable */
  logoUrl?: string | null;
  announcementText: string;
  announcementEnabled: boolean;
  /** @nullable */
  offerBadgeText?: string | null;
  /** @nullable */
  heroTitle?: string | null;
  /** @nullable */
  heroSubtitle?: string | null;
  /** @nullable */
  primaryColor?: string | null;
  /** @nullable */
  footerText?: string | null;
  /** @nullable */
  socialInstagram?: string | null;
  /** @nullable */
  socialFacebook?: string | null;
}

export interface SettingsInput {
  storeName?: string;
  /** @nullable */
  storeTagline?: string | null;
  whatsappNumber?: string;
  /** @nullable */
  logoUrl?: string | null;
  announcementText?: string;
  announcementEnabled?: boolean;
  /** @nullable */
  offerBadgeText?: string | null;
  /** @nullable */
  heroTitle?: string | null;
  /** @nullable */
  heroSubtitle?: string | null;
  /** @nullable */
  primaryColor?: string | null;
  /** @nullable */
  footerText?: string | null;
  /** @nullable */
  socialInstagram?: string | null;
  /** @nullable */
  socialFacebook?: string | null;
}

export type ListProductsParams = {
/**
 * @nullable
 */
category?: string | null;
/**
 * @nullable
 */
search?: string | null;
/**
 * @nullable
 */
featured?: boolean | null;
/**
 * @nullable
 */
inStock?: boolean | null;
};

export type ListOrdersParams = {
/**
 * @nullable
 */
status?: string | null;
/**
 * @nullable
 */
paymentMethod?: string | null;
/**
 * @nullable
 */
search?: string | null;
};