# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ShopEasy e-commerce API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: adminLogin
      tags: [auth]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
...
  /auth/me:
    get:
      operationId: getAuthMe
      tags: [auth]
      summary: Get current admin session
      responses:
        "200":
          description: Admin session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Not authenticated

  # ─── CATEGORIES ─────────────────────────────────────────────────────────────
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create category (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    put:
      operationId: updateCategory
      tags: [categories]
      summary: Update category (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete category (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── PRODUCTS ────────────────────────────────────────────────────────────────
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders (admin)
      parameters:
        - name: status
          in: query
          schema:
            type: ["string", "null"]
        - name: paymentMethod
          in: query
          schema:
            type: ["string", "null"]
        - name: search
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Orders list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Place a new order (customer)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Order placed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
...
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found
  /orders/{id}/status:
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  # ─── BANNERS ─────────────────────────────────────────────────────────────────
  /banners:
    get:
      operationId: listBanners
      tags: [banners]
      summary: List homepage banners
      responses:
        "200":
          description: Banners
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Banner"
    post:
      operationId: createBanner
      tags: [banners]
      summary: Create banner (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BannerInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Banner"
  /banners/{id}:
    put:
      operationId: updateBanner
      tags: [banners]
      summary: Update banner (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BannerInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Banner"
    delete:
      operationId: deleteBanner
      tags: [banners]
      summary: Delete banner (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── SETTINGS ────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    LoginInput:
      type: object
      required: [password]
      properties:
        password:
          type: string
    AuthResponse:
      type: object
      required: [authenticated]
      properties:
        authenticated:
          type: boolean

    # ─── Category ────────────────────────────────────────────────────────────
    Category:
      type: object
      required: [id, name, slug]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        icon:
          type: ["string", "null"]
        productCount:
          type: ["integer", "null"]
    CategoryInput:
      type: object
      required: [name, slug]
      properties:
        name:
          type: string
        slug:
          type: string
        icon:
          type: ["string", "null"]

    # ─── Product ─────────────────────────────────────────────────────────────
...
    OrderItem:
      type: object
      required: [productId, productName, quantity, price, offerPrice]
      properties:
        productId:
          type: integer
        productName:
          type: string
        productImage:
          type: ["string", "null"]
        quantity:
          type: integer
        price:
          type: number
        offerPrice:
          type: number
    Order:
      type: object
      required: [id, orderId, customerName, phone, address, items, total, paymentMethod, status, createdAt]
      properties:
        id:
          type: integer
        orderId:
          type: string
        customerName:
          type: string
        phone:
          type: string
        address:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        total:
          type: number
        paymentMethod:
          type: string
        status:
          type: string
        createdAt:
          type: string
    OrderInput:
      type: object
      required: [customerName, phone, address, items, total, paymentMethod]
      properties:
        customerName:
          type: string
        phone:
          type: string
        address:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        total:
          type: number
        paymentMethod:
          type: string
    OrderStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
...
    Banner:
      type: object
      required: [id, title, imageUrl, active, sortOrder]
      properties:
        id:
          type: integer
        title:
          type: string
        subtitle:
          type: ["string", "null"]
        imageUrl:
          type: string
        linkUrl:
          type: ["string", "null"]
        active:
          type: boolean
        sortOrder:
          type: integer
    BannerInput:
      type: object
      required: [title, imageUrl]
      properties:
        title:
          type: string
        subtitle:
          type: ["string", "null"]
        imageUrl:
          type: string
        linkUrl:
          type: ["string", "null"]
        active:
          type: boolean
        sortOrder:
          type: integer

    # ─── Settings ────────────────────────────────────────────────────────────
...